Don't try to run tests from Maven
Because the tests are run from the ca.ubc.stat.blang subproject,
they don't know about the mock test classes that are defined in
ca.ubc.stat.blang.tests and fail with ClassNotFound exceptions.
So let's use gradle for testing and maven for update site building,
at least for now.